(#15)(#16) Update gulp dependencies and code block
Dependencies that are used in gulpfile.js have been updated to use the
recommended packages. For logging errors, choco-theme was relying on a
depreciated package `gulp-util`. This package has been replaced with
the recommended package `fancy-log`.

This also updates code blocks so that they will render correctly in
dark mode.