(#22)(#24) Remove Gulp & Upgrade Bootstrap 5.3.x
This removes Gulp and introduces a number of new
technologies. The build process is now handled by
esbuild, SASS, and a number of new technologies.
Instead of running `gulp` to compile assets, the
command `yarn run choco-theme` is now used. By
running this, all necessary files are put in place
and CSS is purged.

This upgrades Bootstrap to 5.3.x. A large effort
was made to update choco-theme to use Bootstraps
css variables and dark mode settings. This will
help future upgrades to Bootstrap and choco-theme.